refactor: drop permissions api & check permissions client side
the permissions api is a thin layer checking a wallets permission and caching the result.
while this is nice to reduce rpc calls, it complicates the setup as people have to run a permissions api somewhere.
this pr moves the permissions check to the client, removing the dependency on a elsewhere deployed api.